+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * fix: Allow initialize in other scopes (#229)
 * feat: Rename command `check-eask` to `analyze` (#230)
 * feat(fmt): Add formatters (#232)
+* feat(test): Add built-in `ecukes` test command (#236)
 
 ## 0.9.x
 > Released Nov 17, 2023